Commited some old stuff on my computer:
- a 64-bit clean (I Hope :-) ) version of NewGen. Should work unchanged on
  32 bits. Basically, many ptr-size int types are used, printf formats are
  adapted, some clean up in the signness. But it needs further testing and
  validation. Open question: are NewGen data files compatible between 32
  and 64 bits? If so, on which hypothesis?
 
- a Doxygen configuration.

Next stop: a 64-bit clean linear (I need some help to clean up many
unrelated warning, by the way), and then a 64-bit version of PIPS! :-)


git-svn-id: http://svn.cri.ensmp.fr/svn/newgen/trunk@896 80f2d23d-f6f1-0310-8ea7-b0e6515bd809

+++ src/genC/genC.h
@@ -28,6 +28,7 @@
  */
 /* #include <sys/stdtypes.h> */
 #include <stdio.h>
+#include <stdint.h>
 #include <malloc.h>
 #include <strings.h>
 /* extern char * strdup(const char *);*/
@@ -58,7 +59,7 @@ typedef union gen_chunk {
   unit u;
   bool b;
   char c;
-  int i;
+  intptr_t i;
   float f;
   string s;
   void * e; /* for externals */
